Extract mapping methods to utils file
